plugins/softer_world/pyproject.toml
[project]
name = "softer-world-generator"
version = "0.1.0"
description = "Softer World comic generator with weather integration"
authors = [
    {name = "Joe", email = "joe@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "qrcode[pil]>=7.0.0",
    "font-roboto>=0.0.1",
    "Pillow>=9.0.0",
    "beautifulsoup4>=4.11.0",
    "requests>=2.28.0",
    "flask>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
